fix backwards direction keypress bug
Pressing a key to turn the snake left or right and then immediately pressing
the key to turn backwards kills the snake by moving it backwards.
There is a conditional to prevent this from happening once, but it fails
in the case mentioned before.
This fixes that by saving the direction the player intends to turn, but only
actually taking the direction once we're ready to render the move.